+++ src/ShaderPlayground.Core/CompilerNames.cs
@@ -9,6 +9,7 @@
         public const string Mali = "mali";
         public const string Glslang = "glslang";
         public const string IntelShaderAnalyzer = "intelshaderanalyzer";
+        public const string Metal = "metal";
         public const string Naga = "naga";
         public const string PowerVR = "powervr";
         public const string Rga = "rga";